+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.18.0 (2024-xx-xx)
+## 1.18.0 (2024-02-29)
 
 - Improvement: if a message consists of several photos, now they will all be deleted using the /warn command
 - Fix: displaying the name of the chat with the bot in the settings